requirements.txt
setuptools
pyyaml
wheel
pyinstaller
boto3
awscli
JSON-log-formatter
git+https://github.com/OT-PYTHON-UTILS/ot-aws-libs.git@develop
git+https://github.com/aquasecurity/trivy.git@main